Clean up compiler warnings

+++ src/main.cpp
@@ -6,8 +6,8 @@ int main() {
     simplefluid::SimpleFluid fluid;
     InputHandler input_handler;
 
-    unsigned int window_width = std::ceil(simplefluid::GRID_WIDTH * simplefluid::PIXELS_PER_UNIT);
-    unsigned int window_height = std::ceil(simplefluid::GRID_HEIGHT * simplefluid::PIXELS_PER_UNIT);
+    unsigned int window_width = (unsigned int)std::ceil(simplefluid::GRID_WIDTH * simplefluid::PIXELS_PER_UNIT);
+    unsigned int window_height = (unsigned int)std::ceil(simplefluid::GRID_HEIGHT * simplefluid::PIXELS_PER_UNIT);
     sf::RenderWindow window(sf::VideoMode(window_width, window_height), "Simple Fluid");
 
     sf::Clock clock;
